requirements.txt

numpy
requests
websocket-client
pyopencl
torch
pytest
pandas
